feat(error-handling): restructure error handling module with enhanced types and sanitization
