+++ backend/src/error/mod.rs
@@ -1,0 +1,9 @@
+pub mod code;
+pub mod conversions;
+pub mod response;
+pub mod sanitize;
+pub mod types;
+
+pub use code::ErrorCode;
+pub use response::ErrorResponse;
+pub use types::AppError;
